se volvio al decimal

+++ src/main/java/shareddtos/billmodule/BillItem/BillItemDTO.java
@@ -1,6 +1,5 @@
 package shareddtos.billmodule.BillItem;
 
-import java.math.BigDecimal;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,6 +13,6 @@ public class BillItemDTO {
     private Long productId;
     private String name;
     private int requestedQuantity;
-    private BigDecimal price;
-    private BigDecimal subTotal;
+    private Double price;
+    private Double subTotal;
 }
