+++ src/main/java/shareddtos/billmodule/BillItem/CreateBillItemDTO.java
@@ -1,7 +1,5 @@
 package shareddtos.billmodule.BillItem;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,7 +12,7 @@ public class CreateBillItemDTO {
     private Long productId;
     private String name;
     private int requestedQuantity;
-    private BigDecimal price;
-    private BigDecimal subTotal;
+    private Double price;
+    private Double subTotal;
 
 }
